support: Move chacha20 to support
Relates to #655
