+++ pyatv/companion/srp.py
@@ -27,9 +27,8 @@ from cryptography.hazmat.primitives.asymmetric.x25519 import (
 
 from pyatv import exceptions
 from pyatv.companion import opack
-from pyatv.support import log_binary, hap_tlv8
+from pyatv.support import chacha20, log_binary, hap_tlv8
 from pyatv.support.hap_tlv8 import TlvValue
-from pyatv.mrp import chacha20
 
 _LOGGER = logging.getLogger(__name__)
